move setChannel to support and adjust add modules
